+++ reflect.go
@@ -114,8 +114,11 @@ type ConcreteType struct {
 	Byte byte
 }
 
-// Must use this to register an interface to properly decode the
-// underlying concrete type.
+// This function should be used to register the receiving interface that will
+// be used to decode an underlying concrete type. The interface MUST be embedded
+// in a struct, and the interface MUST be the only field and it MUST be exported.
+// For example:
+//      RegisterInterface(struct{ Animal }{}, ConcreteType{&foo, 0x01})
 func RegisterInterface(o interface{}, ctypes ...ConcreteType) *TypeInfo {
 	it := GetTypeFromStructDeclaration(o)
 	if it.Kind() != reflect.Interface {