+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 0.7.0 (Oct 26, 2017)
+
+BREAKING CHANGE:
+ - time: panic on encode, error on decode for times before 1970
+ - rm codec.go
+
+IMPROVEMENTS:
+ - various additional comments, guards, and checks
+
+BUG FIXES:
+ - fix default encoding of time and bytes
+ - don't panic on ReadTime
+ - limit the amount of memory that can be allocated
+
 ## 0.6.2 (May 18, 2017)
 
 FEATURES: